Move source files to more appropriate directories
