+++ src/sentry/processing/view_hierarchy.cpp
