+++ src/sentry/sentry_options.h
@@ -4,7 +4,7 @@
 #include "sentry/godot_error_types.h"
 #include "sentry/level.h"
 #include "sentry/processing/sentry_event_processor.h"
-#include "sentry/simple_bind.h"
+#include "sentry/util/simple_bind.h"
 
 #include <godot_cpp/classes/ref_counted.hpp>
 #include <godot_cpp/variant/variant.hpp>
