+++ src/sentry/util/simple_bind.h
